+++ examples/GrpcService/Services/GreeterService.cs
@@ -7,13 +7,6 @@ namespace Examples.GrpcService;
 
 public class GreeterService : Greeter.GreeterBase
 {
-    private readonly ILogger<GreeterService> logger;
-
-    public GreeterService(ILogger<GreeterService> logger)
-    {
-        this.logger = logger;
-    }
-
     public override Task<HelloReply> SayHello(HelloRequest request, ServerCallContext context)
     {
         return Task.FromResult(new HelloReply