+++ src/OpenTelemetry/Metrics/ExportModesAttribute.cs
@@ -9,19 +9,17 @@ namespace OpenTelemetry.Metrics;
 [AttributeUsage(AttributeTargets.Class, AllowMultiple = false, Inherited = true)]
 public sealed class ExportModesAttribute : Attribute
 {
-    private readonly ExportModes supportedExportModes;
-
     /// <summary>
     /// Initializes a new instance of the <see cref="ExportModesAttribute"/> class.
     /// </summary>
     /// <param name="supported"><see cref="ExportModes"/>.</param>
     public ExportModesAttribute(ExportModes supported)
     {
-        this.supportedExportModes = supported;
+        this.Supported = supported;
     }
 
     /// <summary>
     /// Gets the supported <see cref="ExportModes"/>.
     /// </summary>
-    public ExportModes Supported => this.supportedExportModes;
+    public ExportModes Supported { get; }
 }