+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/CHANGELOG.md
@@ -7,6 +7,16 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+* Updated OpenTelemetry core component version(s) to `1.11.1`.
+  ([#6088](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6088))
+
+## 1.11.0-beta.1
+
+Released 2025-Jan-16
+
+* Updated OpenTelemetry core component version(s) to `1.11.0`.
+  ([#6064](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6064))
+
 ## 1.10.0-beta.1
 
 Released 2024-Nov-12